OM-14 refactor: FeedCard에서 getTimeAgo 함수를 utils/date.js로 이동
